rename to mock_bm_os.h
