+++ test/stubs/bm_os.c
@@ -1,4 +1,4 @@
-#include "bm_os.h"
+#include "mock_bm_os.h"
 
 DEFINE_FFF_GLOBALS;
